+++ packages/login_client/CHANGELOG.md
@@ -1,6 +1,6 @@
 # 3.2.0
 
-- Always set `tokenEndpoint` using value of `authorizationUri` from `OAuthSettings` during `LoginClient` initialization.
+- Always use `OAuthSettings.authorizationUri` for token endpoint in `LoginClient` initialization (for refreshing the token).
 - Bump `leancode_lint` dev dependency to `12.0.0`.
 - Bump `custom_lint` dev dependency to `0.6.4`.