+++ packages/login_client/CHANGELOG.md
@@ -1,5 +1,6 @@
-# Unreleased
+# 3.2.0
 
+- Always use `OAuthSettings.authorizationUri` for token endpoint in `LoginClient` initialization (for refreshing the token).
 - Bump `leancode_lint` dev dependency to `12.0.0`.
 - Bump `custom_lint` dev dependency to `0.6.4`.